+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 - **[UPDATE]** Migrate `MediaSection` to new mdx story format
+- **[UPDATE]** Migrate `MediaContentSection` to new mdx story format
+- **[UPDATE]** Migrate `IllustratedSection` to new mdx story format
 - **[FIX]** Fixed `MediaContentSection` margins on the title
 [...]